+++ game/src/item.cpp
@@ -34,8 +34,8 @@ Item Item::createItem(uint16_t id, bool isBlock, int tx, int ty) {
     Item res;
     res.Id = id;
     res._isBlock = isBlock;
-    res.tx = tx * texSize / (float)atlasWidth;
-    res.ty = ty * texSize / (float)atlasHeight;
+    res.tx = (float)(tx * texSize) / (float)atlasWidth;
+    res.ty = (float)(ty * texSize) / (float)atlasHeight;
     res.txi = tx;
     res.tyi = ty;
     return res;
